Have client send a request
